group components index properties
